Implemented SMULL, SMLAL, UMULL and UMLAL in the interpreter.

+++ src/Common/MathHelper.cpp
@@ -10,7 +10,7 @@ bool MathHelper::CheckBits(uint32_t value, uint8_t start, uint8_t bitCount, uint
     return ((value >> start) & ((1 << bitCount) - 1)) == check;
 }
 
-uint32_t MathHelper::GetBits(uint32_t value, uint8_t start, uint8_t bitCount)
+uint32_t MathHelper::GetBits(uint64_t value, uint8_t start, uint8_t bitCount)
 {
     return ((value >> start) & ((1 << bitCount) - 1));
 }
